+++ pakyow-core/lib/pakyow/plugin.rb
@@ -62,10 +62,8 @@ module Pakyow
 
     configurable :tasks do
       setting :prelaunch, []
-    end
 
-    configurable :commands do
-      setting :prelaunch, []
+      deprecate :prelaunch
     end
 
     include Support::Definable