+++ pakyow-core/spec/features/commands/help_spec.rb
@@ -19,7 +19,7 @@ RSpec.describe "cli: help" do
 
   describe "running" do
     it "displays help for the cli" do
-      expect(run_command(command, project: true)).to eq("\e[34;1mPakyow Command Line Interface\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow [COMMAND]\n\n\e[1mCOMMANDS\e[0m\n  boot            \e[33mBoot the project\e[0m\n  help            \e[33mGet help for the command line interface\e[0m\n  info            \e[33mShow details about the current project\e[0m\n  info:endpoints  \e[33mShow defined endpoints for an app\e[0m\n  irb             \e[33mStart an interactive session\e[0m\n  prelaunch       \e[33mRun the prelaunch commands\e[0m\n  prototype       \e[33mBoot the prototype\e[0m\n")
+      expect(run_command(command, project: true)).to eq("\e[34;1mPakyow Command Line Interface\e[0m\n\n\e[1mUSAGE\e[0m\n  $ pakyow [COMMAND]\n\n\e[1mCOMMANDS\e[0m\n  boot               \e[33mBoot the project\e[0m\n  help               \e[33mGet help for the command line interface\e[0m\n  info               \e[33mShow details about the current project\e[0m\n  info:endpoints     \e[33mShow defined endpoints for an app\e[0m\n  irb                \e[33mStart an interactive session\e[0m\n  prelaunch          \e[33mRun all phases of the prelaunch sequence\e[0m\n  prelaunch:build    \e[33mRun the build phase of the prelaunch sequence\e[0m\n  prelaunch:release  \e[33mRun the release phase of the prelaunch sequence\e[0m\n  prototype          \e[33mBoot the prototype\e[0m\n")
     end
 
     it "displays help for a command" do