+++ pakyow-data/lib/pakyow/commands/db/migrate.rb
@@ -6,6 +6,16 @@ command :db, :migrate, boot: false do
   option :adapter, "The database adapter", default: -> { Pakyow.config.data.default_adapter }
   option :connection, "The database connection", default: -> { Pakyow.config.data.default_connection }
 
+  prelaunch :release do |command|
+    Pakyow.data_connections.each do |adapter, connections|
+      connections.each do |connection_name, connection|
+        if connection.migratable? && connection_name != :memory
+          command.call(adapter: adapter, connection: connection_name)
+        end
+      end
+    end
+  end
+
   action do
     require "pakyow/data/migrator"