+++ pakyow-data/spec/features/prelaunch_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe "data prelaunch tasks" do
+RSpec.describe "data prelaunch commands" do
   before do
     Pakyow.configure do
       Pakyow.config.data.connections.sql[:default] = ENV["DATABASE_URL__POSTGRES"]
@@ -9,12 +9,26 @@ RSpec.describe "data prelaunch tasks" do
 
   include_context "app"
 
-  it "registers a prelaunch command for each migratable connection" do
-    expect(Pakyow.config.commands.prelaunch).to include(["db:migrate", { adapter: :sql, connection: :default }])
-    expect(Pakyow.config.commands.prelaunch).to include(["db:migrate", { adapter: :sql, connection: :another }])
-  end
+  describe "db:migrate" do
+    it "is a prelaunch command" do
+      expect(Pakyow.command(:db, :migrate).prelaunch?).to be(true)
+    end
 
-  it "does not register in-memory connections" do
-    expect(Pakyow.config.commands.prelaunch).not_to include(["db:migrate", { adapter: :sql, connection: :memory }])
+    it "is part of the release phase" do
+      expect(Pakyow.command(:db, :migrate).prelaunch_phase).to eq(:release)
+    end
+
+    it "yields each migratable connection" do
+      prelaunches = []
+
+      Pakyow.command(:db, :migrate).prelaunches do |**args|
+        prelaunches << args
+      end
+
+      expect(prelaunches).to eq([
+        { adapter: :sql, connection: :default },
+        { adapter: :sql, connection: :another }
+      ])
+    end
   end
 end