+++ pakyow-core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # v1.1.0 (unreleased)
 
+  * `add` **Split prelaunch commands into `prelaunch:build` and `prelaunch:release` phases.**
+    - Build phase commands don't require a full boot and emit artifacts that alter the environment,
+      such as by precompiling assets to the filesystem or uploading them to a CDN. The build phase
+      runs when the project is built.
+    - Release phase commands require a full boot and have the ability to directly alter the
+      environment, such as by running database migrations. The release phase runs within the
+      deployed project just before it boots.
+
+    *Related links:*
+    - [Pull Request #416][pr-416]
+
   * `chg` **Handle defining the same app multiple times.**
 
     *Related links:*
@@ -282,6 +293,7 @@
     *Related links:*
     - [Pull Request #338][pr-338]
 
+[pr-416]: https://github.com/pakyow/pakyow/pull/416
 [pr-413]: https://github.com/pakyow/pakyow/pull/413
 [pr-412]: https://github.com/pakyow/pakyow/pull/412
 [pr-411]: https://github.com/pakyow/pakyow/pull/411